pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "The site analyzes the specified pages for SEO suitability"
authors = ["Garkusha Valery <aston585v@gmail.com>"]
license = "Freeware"
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.0"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
postgres = "^4.0"
psycopg2 = "^2.9.9"
validators = "^0.22.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
autopep8 = "^2.0.4"
pytest = "^7.4.3"









[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
